refactor: moved entity-state tests to separate directory
