+++ src/tests/entity-state/static-selectors.spec.ts
@@ -1,8 +1,8 @@
 import { State } from '@ngxs/store';
-import { defaultEntityState, EntityState } from '../lib/entity-state';
-import { EntityStateModel } from '../lib/models';
-import { IdStrategy } from '../lib/id-strategy';
-import { NGXS_META_KEY } from '../lib/internal';
+import { defaultEntityState, EntityState } from '../../lib/entity-state';
+import { EntityStateModel } from '../../lib/models';
+import { IdStrategy } from '../../lib/id-strategy';
+import { NGXS_META_KEY } from '../../lib/internal';
 
 interface ToDo {
   title: string;
